enhancement: refactored and extracted code cell
