+++ src/index.tsx
@@ -1,33 +1,11 @@
 import "bulmaswatch/superhero/bulmaswatch.min.css";
 import ReactDOM from "react-dom";
-import { useState } from "react";
-import CodeEditor from "./components/code-editor";
-import Preview from "./components/preview";
-import bundle from "./bundler";
+import CodeCell from "./components/code-cell";
 
 const App = () => {
-  const [code, setCode] = useState("");
-  const [input, setInput] = useState("");
-
-  const onClick = async () => {
-    const output = await bundle(input);
-    setCode(output);
-  };
-
   return (
     <div>
-      <CodeEditor
-        initialValue="const a = 1;"
-        onChange={(value) => setInput(value)}
-      />
-      <textarea
-        value={input}
-        onChange={(e) => setInput(e.target.value)}
-      ></textarea>
-      <div>
-        <button onClick={onClick}>Submit</button>
-      </div>
-      <Preview code={code} />
+      <CodeCell />
     </div>
   );
 };
